Create AppState and store application current state there rather than modify Namespace
