+++ src/pyzet/constants.py
@@ -7,7 +7,9 @@ from pathlib import Path
 VERSION = metadata.version('pyzet')
 
 CONFIG_FILE = 'pyzet.yaml'
-DEFAULT_CFG_LOCATION = Path(Path.home(), '.config', 'pyzet', CONFIG_FILE)
+DEFAULT_CFG_LOCATION = Path(
+    Path.home(), '.config', 'pyzet', CONFIG_FILE
+).as_posix()
 
 DEFAULT_BRANCH = 'main'
 DEFAULT_REMOTE_NAME = 'origin'
