+++ src/pyzet/main.py
@@ -1,14 +1,10 @@
 from __future__ import annotations
 
-import argparse
-from argparse import ArgumentParser
-from argparse import Namespace
-
-import pyzet.constants as C
+import pyzet.cli
 from pyzet import config
 from pyzet import show
 from pyzet import utils
-from pyzet.grep import define_grep_cli
+from pyzet.cli import get_parser
 from pyzet.grep import grep
 from pyzet.ops import add_zettel
 from pyzet.ops import clean_zet_repo
@@ -20,18 +16,17 @@ from pyzet.ops import init_repo
 from pyzet.ops import list_tags
 from pyzet.ops import list_zettels
 from pyzet.ops import remove_zettel
-from pyzet.sample_config import define_sample_config_cli
 from pyzet.sample_config import sample_config
-from pyzet.utils import add_pattern_args
 from pyzet.utils import call_git
 
 
 def main(argv: list[str] | None = None) -> int:
     utils.configure_console_print_utf8()
 
-    parser, subparsers = _get_parser()
-    args = parser.parse_args(argv)
-    utils.setup_logger(utils.compute_log_level(args.verbose))
+    parser, subparsers = get_parser()
+    args_cli = parser.parse_args(argv)
+    utils.setup_logger(utils.compute_log_level(args_cli.verbose))
+    args = pyzet.cli.populate_args(args_cli, parser)
 
     if args.command == 'show' and args.show_cmd is None:
         subparsers['show'].print_usage()
@@ -45,163 +40,7 @@ def main(argv: list[str] | None = None) -> int:
         raise SystemExit
 
 
-def _get_parser() -> tuple[ArgumentParser, dict[str, ArgumentParser]]:
-    parser = argparse.ArgumentParser(
-        prog='pyzet', formatter_class=argparse.RawTextHelpFormatter
-    )
-    subparsers_dict = {}
-
-    parser.add_argument('-r', '--repo', help='point to a custom ZK repo')
-    parser.add_argument(
-        '-c',
-        '--config',
-        default=C.DEFAULT_CFG_LOCATION,
-        help='which config file to use (default: %(default)s)',
-    )
-    parser.add_argument(
-        '-V',
-        '--version',
-        action='version',
-        version=f'%(prog)s {C.VERSION}',
-    )
-    parser.add_argument(
-        '-v',
-        '--verbose',
-        action='count',
-        default=0,
-        help='increase verbosity of the output',
-    )
-
-    subparsers = parser.add_subparsers(dest='command')
-
-    init_parser = subparsers.add_parser(
-        'init', help='initialize a git ZK repo at configured or given path'
-    )
-    init_parser.add_argument(
-        'path',
-        nargs='?',
-        help='use a different dir than specified in the config file;'
-        ' if target dir exists, it must be empty',
-    )
-    init_parser.add_argument(
-        '-b',
-        '--initial-branch',
-        default=C.DEFAULT_BRANCH,
-        help='initial branch name (default: %(default)s)',
-    )
-
-    subparsers.add_parser('add', help='add a new zettel')
-
-    edit_parser = subparsers.add_parser('edit', help='edit an existing zettel')
-    add_pattern_args(edit_parser)
-
-    remove_parser = subparsers.add_parser('rm', help='remove a zettel')
-    add_pattern_args(remove_parser)
-
-    subparsers_dict['show'] = show.get_parser(subparsers)
-
-    print_parser = subparsers.add_parser(
-        'print',
-        help="print zettel contents, a shorthand for 'pyzet show text'",
-    )
-    add_pattern_args(print_parser)
-
-    list_parser = subparsers.add_parser('list', help='list all zettels')
-    list_parser.add_argument(
-        '-p',
-        '--pretty',
-        action='store_true',
-        help='use prettier format for printing date and time',
-    )
-    list_parser.add_argument(
-        '--tags',
-        action='store_true',
-        help='show tags for each zettel',
-    )
-    list_parser.add_argument(
-        '-l',
-        '--link',
-        action='store_true',
-        help='print zettels as relative Markdown'
-        ' links for pasting in other zettels',
-    )
-    list_parser.add_argument(
-        '-r',
-        '--reverse',
-        action='store_true',
-        help='reverse the output (newest first)',
-    )
-
-    tags_parser = subparsers.add_parser(
-        'tags', help='list all tags and count them'
-    )
-    tags_parser.add_argument(
-        '-r',
-        '--reverse',
-        action='store_true',
-        help='reverse the output',
-    )
-
-    clean_parser = subparsers.add_parser(
-        'clean', help=f"delete empty folders in '{C.ZETDIR}' folder in ZK repo"
-    )
-    clean_parser.add_argument(
-        '-d',
-        '--dry-run',
-        action='store_true',
-        help="list what will be deleted, but don't delete it",
-    )
-    clean_parser.add_argument(
-        '-f',
-        '--force',
-        action='store_true',
-        help='force, use it to actually delete anything',
-    )
-
-    define_grep_cli(subparsers)
-
-    status_parser = subparsers.add_parser(
-        'status', help="run 'git status' in ZK repo"
-    )
-    _add_git_cmd_options(status_parser, 'status')
-
-    subparsers.add_parser('pull', help="run 'git pull --rebase' in ZK repo")
-
-    push_parser = subparsers.add_parser(
-        'push', help="run 'git push' in ZK repo"
-    )
-    _add_git_cmd_options(push_parser, 'push')
-
-    remote_parser = subparsers.add_parser(
-        'remote', help="run 'git remote get-url' in ZK repo"
-    )
-    remote_parser.add_argument(
-        '--name',
-        default=C.DEFAULT_REMOTE_NAME,
-        help='name of git repo remote (default: %(default)s)',
-    )
-
-    subparsers.add_parser('info', help='show stats about ZK repo')
-
-    subparsers.add_parser('query', help='query ZK repo')
-
-    define_sample_config_cli(subparsers)
-
-    return parser, subparsers_dict
-
-
-def _add_git_cmd_options(parser: ArgumentParser, cmd_name: str) -> None:
-    parser.add_argument(
-        'options',
-        action='store',
-        type=str,
-        nargs='*',
-        default=[],
-        help=f"pass 'git {cmd_name}' options, use '--' before including flags",
-    )
-
-
-def _parse_args(args: Namespace) -> int:
+def _parse_args(args: pyzet.cli.AppState) -> int:
     if args.command == 'sample-config':
         sample_config(args.kind)
         return 0
